requirements.txt
streamlit==1.50.0
google-cloud-bigquery==3.40.1
pandas==2.3.3
requests==2.32.5
db-dtypes==1.5.0
streamlit-searchbox==0.1.23
